Sudo manager
The sudo manager has two purposes.  One is to construct an initial
command that is already prefixed with sudo (if necessary).  It also
starts up a background thread to periodically invoke "sudo -v" so that
the user isn't prompted for the sudo password in the middle of an
operation.

The tests have a minor flaw.  Because the tests are run in parallel,
it ends up invoking sudo -v multiple times, which presents confusing
prompts to the user.  This can be remediated by running "sudo -v" from
the command line before running the test.