Framework and start of lvm
This is an initial skeleton of the rback code.  There is a config file
reader, and an interface to gethostname().  These are currently
untested.

The main part of this contribution is the lvm module, which reads the
output of the `lvs` command and gathers the names of each logical
volume on the system.