Implement 'clone' command
The clone command compares a source zfs volume tree, and a mirror tree
and zfs send/recv transfers any missing snapshots to the mirror tree.
Attributes are not copied (not done with -R on send).  This code
currently requires the destination volumes to exist (and they should
be created with any necessary attributes set, such as compression and
xattr/posixacl).