Add the bksure command
'rsure' now supports storing sure data in a BitKeeper archive (which
saves significant space).  Add a new command 'bksure' which stores
sures of the snapshots in to this archive instead of individual files.
The store is expected to be in 'base/bksure' instead of 'base/sure'.
The directory should be created with 'bknew base/bksure'.