Generalize zfs clone code
Move the 'src' and 'dest' values into a new struct, parameterized by a
`ZfsPath`.  Right now, the only implementation is for &str, which
replicates the current behavior.  However, instead of just
constructing a Command directly with the 'zfs' command, this trait
allows this to be changed, for example to use ssh to access data on a
remote host.