Allow clone to empty destination
It isn't quite perfect.  The destination volumes must exist, but must
be empty (at least have no snapshots, but all files will be lost in
this case).

To make this work, create the necessary volumes on the destination,
and set any attributes that are needed (especiallyl things like
acltype that affect how acls are stored).

It isn't completely robus, and sometimes the receive fails with an
error about being unable to unmount the filesystem.  It then hangs,
sending data nowhere while the system just sits there.  You can abort
(with Ctrl-C) and restart, however, and it will continue where it left
off.